% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_state.R
\name{compute_panel_state}
\alias{compute_panel_state}
\title{Title}
\usage{
compute_panel_state(data, scales, keep_state = NULL, drop_state = NULL)
}
\arguments{
\item{county}{}
}
\description{
Title
}
\examples{
library(dplyr)
#state_attributes |> rename(state = state_name) |> compute_panel_state() |> head()
}
