% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_state.R
\name{geom_state}
\alias{geom_state}
\title{Title}
\usage{
geom_state(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  crs = 4326,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(state_attributes) +
aes(state = state_name) +
geom_state()

ggplot(state_attributes) +
aes(state = state_name) +
geom_state(drop_state = c("Hawaii", "Alaska"))

ggplot(state_attributes) +
aes(state = state_name) +
geom_state(drop_state = c("Hawaii", "Alaska")) +
aes(fill = Frost)
}
